{cgan,gen}: Rename cgan -> gen
